addons/ventilairsec_enocean/requirements.txt
pyserial==3.5
requests==2.31.0
cryptography==41.0.7
